requirements.txt
# 股票量化交易系统依赖
pandas>=1.5.0
numpy>=1.21.0
pyyaml>=6.0
yfinance>=0.2.0
requests>=2.28.0
